Website/requirements.txt
arrow==1.3.0
asgiref==3.8.1
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-mgmt-containerregistry==14.0.0
azure-mgmt-core==1.5.0
azure-servicebus==7.12.0
azure-storage-blob==12.25.1
backports.tarfile==1.2.0
binaryornot==0.4.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
cryptography==44.0.2
Django==5.1.6
django-browser-reload==1.18.0
django-ckeditor-5==0.2.18
django-environ==0.12.0
django-extensions==3.2.3
django-js-asset==3.1.2
django-oauth-toolkit==3.0.1
django-taggit==6.1.0
django-tailwind==3.8.0
django-webpack-loader==3.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
docker==7.1.0
dotenv==0.9.9
Flask==3.1.0
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-artifact-registry==1.15.1
google-cloud-build==3.31.1
google-cloud-container==2.56.0
google-cloud-core==2.4.3
google-cloud-iam==2.18.1
google-cloud-monitoring==2.27.1
google-cloud-run==0.10.16
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
grpc-google-iam-v1==0.14.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jwcrypto==1.5.6
keyring==25.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
msal==1.32.3
msal-extensions==1.3.1
oauthlib==3.2.2
packaging==24.2
pillow==11.3.0
pkce==1.0.3
proto-plus==1.26.1
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.11.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
readchar==4.2.1
requests==2.32.3
rich==13.9.4
rsa==4.9
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise==6.9.0
zipp==3.21.0
pydantic
azure-keyvault-secrets
azure-identity
